# Conflicts:
#	cmake/CompileFlags.cmake
#	src/cargomonitor.cpp
#	src/core/CMakeLists.txt
#	src/economy.cpp
#	src/landscape.cpp
#	src/linkgraph/flowmapper.cpp
#	src/linkgraph/linkgraph_gui.cpp
#	src/linkgraph/linkgraphschedule.cpp
#	src/misc_gui.cpp
#	src/newgrf_generic.cpp
#	src/newgrf_storage.cpp
#	src/rail_gui.cpp
#	src/saveload/afterload.cpp
#	src/saveload/station_sl.cpp
#	src/script/script_gui.cpp
#	src/station_cmd.cpp
#	src/station_gui.cpp
#	src/string_func.h
#	src/terraform_cmd.cpp

+++ src/unit_conversion.h
@@ -17,10 +17,10 @@ uint ConvertWeightToDisplayWeight(uint weight);
 uint ConvertDisplayWeightToWeight(uint weight);
 uint ConvertPowerToDisplayPower(uint power);
 uint ConvertDisplayPowerToPower(uint power);
-uint ConvertForceToDisplayForce(uint force);
-uint ConvertDisplayForceToForce(uint force);
-void ConvertPowerWeightRatioToDisplay(uint ratio, int64 &value, int64 &decimals);
-void ConvertForceWeightRatioToDisplay(uint ratio, int64 &value, int64 &decimals);
+int64 ConvertForceToDisplayForce(int64 force);
+int64 ConvertDisplayForceToForce(int64 force);
+void ConvertPowerWeightRatioToDisplay(int64 ratio, int64 &value, int64 &decimals);
+void ConvertForceWeightRatioToDisplay(int64 ratio, int64 &value, int64 &decimals);
 uint ConvertDisplayToPowerWeightRatio(double in);
 uint ConvertDisplayToForceWeightRatio(double in);
